+++ packages/components/pie-switch/test/helpers/constants.ts
@@ -1,0 +1,1 @@
+export const EXPECTED_EVENT_MESSAGE = 'Switch clicked';